Feat : 구입 금액 입력받기
